+++ test/unit/certificate-list.test.js
@@ -14,7 +14,7 @@ describe('CertificateList', function () {
             matches: ['https://example.com', 'https://*.foo.com']
         }]);
 
-        it('should return a certificate instance when matched, and undefined instead', function() {
+        it('should return a certificate instance when matched, and undefined instead', function () {
             var matchedCertificate = certificateList.resolveOne('https://www.google.com');
             expect(matchedCertificate instanceof Certificate).to.be.ok();
 
@@ -22,19 +22,19 @@ describe('CertificateList', function () {
             expect(matchedCertificate).to.eql(undefined);
         });
 
-        it('should return the matched certificate', function() {
+        it('should return the matched certificate', function () {
             var matchedCertificate = certificateList.resolveOne('https://www.google.com');
 
             expect(matchedCertificate.id).to.eql(1);
         });
 
-        it('should return undefined when no certificate is matched', function() {
+        it('should return undefined when no certificate is matched', function () {
             var matchedCertificate = certificateList.resolveOne('https://www.twitter.com');
 
             expect(matchedCertificate).to.eql(undefined);
         });
 
-        it('should exit safely when called with url which is not a string or a Url', function() {
+        it('should exit safely when called with url which is not a string or a Url', function () {
             var matchedCertificate = certificateList.resolveOne({});
 
             expect(matchedCertificate).to.eql(undefined);
@@ -45,8 +45,8 @@ describe('CertificateList', function () {
         });
     });
 
-    describe('isCertificateList', function() {
-        it('should return true for CertificateList instance', function() {
+    describe('isCertificateList', function () {
+        it('should return true for CertificateList instance', function () {
             var certificateList = new CertificateList({}, [{ matches: [] }]);
 
             expect(CertificateList.isCertificateList()).to.eql(false);