+++ test/unit/property.test.js
@@ -3,6 +3,20 @@ var expect = require('expect.js'),
 
 /* global describe, it */
 describe('Property', function () {
+    describe('sanity', function () {
+        it('initializes successfully', function () {
+            expect((new Property()) instanceof Property).to.be.ok();
+        });
+        it('allows a description to be set', function () {
+            var prop = new Property();
+            expect(prop.describe).to.be.a('function');
+            expect(prop.describe.bind(prop)).withArgs('Hello Postman').to.not.throwException();
+            expect(prop.description).to.be.ok();
+            expect(prop.description.toString()).to.be('Hello Postman');
+            expect(prop.description.type).to.be('text/plain');
+        });
+    });
+
     describe('variable resolution', function () {
         it('must resolve variables accurately', function () {
             var unresolvedRequest = {
@@ -41,7 +55,7 @@ describe('Property', function () {
             expect(resolved).to.eql(expectedResolution);
         });
 
-        it('must resolve variables accurately', function () {
+        it('must resolve variables in complex nested objects correctly', function () {
             var unresolvedRequest = {
                     one: '{{alpha}}-{{beta}}-{{gamma}}',
                     two: { a: '{{alpha}}', b: '{{beta}}', c: '{{gamma}}' },