+++ test/unit/request-body.test.js
@@ -1,7 +1,8 @@
 var expect = require('expect.js'),
     rawRequestBody = require('../fixtures').requestData,
     RequestBody = require('../../lib/index.js').RequestBody,
-    QueryParam = require('../../lib/index.js').QueryParam;
+    QueryParam = require('../../lib/index.js').QueryParam,
+    PropertyList = require('../../lib/index.js').PropertyList;
 
 /* global describe, it */
 describe('RequestBody', function () {
@@ -78,6 +79,25 @@ describe('RequestBody', function () {
         });
     });
 
+    describe('sanity', function () {
+        it('should be parsed properly', function () {
+            var reqData = new RequestBody(rawRequestBody);
+            expect(reqData).to.be.ok();
+            expect(reqData).to.have.property('mode', rawRequestBody.mode);
+
+            // Raw Request body
+            expect(reqData).to.have.property('raw', rawRequestBody.raw);
+
+            // Form data
+            expect(reqData.formdata).to.be.a(PropertyList);
+            expect(reqData.formdata.count()).to.eql(rawRequestBody.formdata.length);
+
+            // URL Encoded parameters
+            expect(reqData.urlencoded).to.be.a(PropertyList);
+            expect(reqData.urlencoded.count()).to.eql(rawRequestBody.urlencoded.length);
+        });
+    });
+
     describe('isEmpty', function () {
         it('should return true if no request body is set', function () {
             var body = new RequestBody();
@@ -146,7 +166,7 @@ describe('RequestBody', function () {
                 mode: 'urlencoded',
                 urlencoded: [{
                     key: 'haha',
-                    value: { some: 'random', javascript: 'object' }  // this functionality is used in the app
+                    value: { some: 'random', javascript: 'object' } // this functionality is used in the app
                 }]
             });
             expect(body.isEmpty()).to.be(false);